Allow custom cost types with individual price change factors
This rewrite introduces breaking changes that are necessary to lift
the limitation of three cost types (demand, other, proceeds) with
their fixed price change factors. Now the user can define their
own cost types, but has to be explicit about their names, contents
and price change factors. This data, wrapped as a dictionary, is
now required as an input to calc_annuities(). It must include the
desired sign to differentiate between costs and proceeds (which was
hardcoded before).

Additionally, this commit introduces calc_amortization() for
calculating the amortization time.